Refactor : AllRecruitmentOfClubResponse 외부 객체로 분리
