+++ src/main/java/com/greedy/mokkoji/api/recruitment/controller/RecruitmentController.java
@@ -3,8 +3,8 @@ package com.greedy.mokkoji.api.recruitment.controller;
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.AuthCredential;
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.Authentication;
 import com.greedy.mokkoji.api.recruitment.dto.request.RecruitmentCreateRequest;
-import com.greedy.mokkoji.api.recruitment.dto.response.AllRecruitmentOfClubResponse;
-import com.greedy.mokkoji.api.recruitment.dto.response.AllRecruitmentResponse;
+import com.greedy.mokkoji.api.recruitment.dto.response.allRecruitmentOfClub.AllRecruitmentOfClubResponse;
+import com.greedy.mokkoji.api.recruitment.dto.response.AllRecruitment.AllRecruitmentResponse;
 import com.greedy.mokkoji.api.recruitment.dto.response.RecruitmentCreateResponse;
 import com.greedy.mokkoji.api.recruitment.dto.response.SpecificRecruitmentResponse;
 import com.greedy.mokkoji.api.recruitment.service.RecruitmentService;
