+++ src/main/java/com/greedy/mokkoji/api/recruitment/dto/request/RecruitmentCreateRequest.java
@@ -1,11 +1,8 @@
 package com.greedy.mokkoji.api.recruitment.dto.request;
 
-import lombok.Builder;
-
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Builder
 public record RecruitmentCreateRequest(
         String title,
         List<String> images,
