+++ src/main/java/com/greedy/mokkoji/api/recruitment/dto/response/AllRecruitment/RecruitmentPreviewResponse.java
@@ -1,0 +1,9 @@
+package com.greedy.mokkoji.api.recruitment.dto.response.AllRecruitment;
+
+import lombok.Builder;
+
+@Builder
+public record RecruitmentPreviewResponse(
+
+) {
+}
